problems/257.二叉树的所有路径/index_test.go
package index

import (
	"reflect"
	"testing"

	"github.com/penguinn/leetcode/common"
)

func Test_binaryTreePaths(t *testing.T) {
	type args struct {
		root *common.TreeNode
	}
	tests := []struct {
		name string
		args args
		want []string
	}{
		{
			name: "test1",
			args: args{root: common.NewBinaryTree([]interface{}{1, 2, 3, nil, 5})},
			want: []string{"1->2->5", "1->3"},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := binaryTreePaths(tt.args.root); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("binaryTreePaths() = %v, want %v", got, tt.want)
			}
		})
	}
}
